Add optional, automatic OOM logger parsing
An updated ESP8266 panic function can print out the calling function/line
and size requested for the last malloc/realloc/calloc/new allocation that
failed, without the overhead of full the OOM stack.

Add parsing for this line, when present, and output the function, file,
line, and amount of memory requested to the display.  When not present,
do nothing different.