Replace addr2line with GDB to get accurate line #s
Addr2line on the xtensa seems to have trouble with identifying the
proper file:line number for an exception address.  Until there is a
fix, the workaround is to use GDB to locate them.

See https://github.com/jcmvbkbc/binutils-gdb-xtensa/issues/5 .

Replace the addr2line processing/formatting with gdb's formatting,
leaving the output identical with one exception:  addresses with
no source code (i.e. not code, but constant data somewhere or just
random variables on the stack) *will not print*.  They're silently
ignored in the output.

This also now presents the EPC1/PC and EXCVADDR on their own
lines hilighted at the top of the dump.

Additionally, handle the case where the exception dumper itself
on the ESP8266 hits the WDT (i.e. during dump of an infinite
recursion).  In this case the final "<<<stack<<<" line may not be
generated, so simply parse everything from ">>>stack>>>" to the end.